pbd client: adds nodes pagination cursor
To make iterating over pages of nodes easier and consistent, I added a
pagination cursor that allows a user of the nodes client to call
NodesCursor.Next() to keep fetching pages until an io.EOF error is
returned. This keeps the client use minimal of implementing packages
(such as scarp) and centralizes and hides the pagination state/logic so
it doesn't need to be implemented over and over again.

I think we should generalize this a bit more and add it to all the pdb
calls we export in this package as clients. I'm not sure if we can
generalize it for the entire go-pe-client tree because other pagination
schemes across PE might be different than pdb's API.