pdb client: correct pagination offset
Here I'm ensuring the pagination object's offset isn't increased if
there is an error when fetching the next page. Before, if the pdb client
returned an error from the puppetdb api, it would keep increasing the
offset, which would not be correct if the server errors ever resolved
and a caller tried calling next on the cursor again.